[IN-155] IN-175 refactor: move GoogleLoginButton component to a separate file
