+++ src/features/auth/api/auth.api.ts
@@ -12,6 +12,10 @@ import { AxiosInstance, AxiosResponse } from 'axios'
 
 class AuthApi {
   constructor(private instance: AxiosInstance) {}
+  public async authWithGoogle(code: string): Promise<AxiosResponse> {
+    return this.instance.post(Endpoints.AuthWithGoogle, { code })
+  }
+
   public async emailConfirmation(
     emailConfirmation: EmailConfirmationRequestData
   ): Promise<AxiosResponse> {
@@ -34,8 +38,12 @@ class AuthApi {
     return instance.post(Endpoints.logout)
   }
 
-  public async me(): Promise<Profile> {
-    return instance.get(Endpoints.me).then((res) => res.data)
+  public async me(accessToken: string): Promise<Profile> {
+    return instance.get(Endpoints.me, {
+      headers: {
+        Authorization: `Bearer ${accessToken}`,
+      },
+    })
   }
 
   public async recoverPassword(
@@ -47,7 +55,6 @@ class AuthApi {
   public async signUp(data: SignUpRequestData): Promise<AxiosResponse> {
     return this.instance.post(Endpoints.SignUp, data)
   }
-
   public async updateToken() {
     return instance
       .post(Endpoints.updateToken)
