+++ src/features/auth/ui/signUp/GoogleLoginButton.tsx
@@ -1,0 +1,52 @@
+import React from 'react'
+
+import { generalStore } from '@/app/store'
+import { GoogleSvgrepoCom1 } from '@/assets/icons/filledIcons'
+import { responseErrorHandler } from '@/common/utils/responseErrorHandler'
+import { authApi } from '@/features/auth'
+import authStore from '@/features/auth/model/authStore'
+import { useGoogleLogin } from '@react-oauth/google'
+import { useRouter } from 'next/router'
+
+export const GoogleLoginButton = React.forwardRef<HTMLButtonElement>(
+  (_, ref) => {
+    const router = useRouter()
+
+    const isLoading = generalStore
+
+    const login = useGoogleLogin({
+      flow: 'auth-code',
+      onError: () => {},
+      onSuccess: async (credentialResponse) => {
+        if (credentialResponse.code) {
+          try {
+            isLoading.turnOnLoading()
+            const res = await authApi.authWithGoogle(credentialResponse.code)
+
+            localStorage.setItem('accessToken', res.data.accessToken)
+            await authStore.me(res.data.accessToken)
+            isLoading.turnOffLoading()
+            await router.push('/profile')
+          } catch (error) {
+            responseErrorHandler(error)
+          } finally {
+            isLoading.turnOffLoading()
+          }
+        }
+      },
+    })
+
+    return (
+      <button
+        onClick={() => login()}
+        ref={ref}
+        type={'button'}
+      >
+        <GoogleSvgrepoCom1
+          height={36}
+          width={36}
+        />
+      </button>
+    )
+  }
+)
